+++ services/api/src/app/messages/messages.controller.spec.ts
@@ -12,11 +12,11 @@ import { MessageFactory } from '@fixtures/messages/message.factory';
 import { UsersRepository } from '@entities/users';
 import { MessagesRepository } from '@entities/messages';
 import { TestDataModule } from '@fixtures/data/test.data.module';
-import { TestUsersRepository } from '@fixtures/data/test.users.repository';
-import { TestMessagesRepository } from '@fixtures/data/test.messages.repository';
+import { TestUsersRepository } from '@data/repositories/test/test.users.repository';
+import { TestMessagesRepository } from '@data/repositories/test/test.messages.repository';
 import { Room } from '@entities/room.entity';
 import { RoomsRepository } from '@entities/rooms.repository';
-import { TestRoomsRepository } from '@fixtures/data/test.rooms.repository';
+import { TestRoomsRepository } from '@data/repositories/test/test.rooms.repository';
 import { RoomFactory } from '@fixtures/messages/room.factory';
 import { CreateMessageDto } from './dto/messages';
 import { TestMembershipsRepository } from '@fixtures/data/test.memberships.repository';
