+++ services/api/src/data/repositories/dynamodb.messages.repository.e2e-spec.ts
@@ -1,6 +1,6 @@
-import { DynamoDBMessagesRepository } from '@data/repositories/dynamodb.messages.repository';
+import { DynamoDBMessagesRepository } from '@data/repositories/dynamodb/dynamodb.messages.repository';
 import { DraftMessage } from '@entities/messages';
-import { TestMessagesRepository } from '@fixtures/data/test.messages.repository';
+import { TestMessagesRepository } from '@data/repositories/test/test.messages.repository';
 import { RoomFactory } from '@fixtures/messages/room.factory';
 import { UserFactory } from '@fixtures/messages/user.factory';
 import { Test, TestingModule } from '@nestjs/testing';
