+++ services/api/src/data/repositories/dynamodb.rooms.repository.e2e-spec.ts
@@ -1,7 +1,7 @@
-import { DynamoDBRoomsRepository } from '@data/repositories/dynamodb.rooms.repository';
+import { DynamoDBRoomsRepository } from '@data/repositories/dynamodb/dynamodb.rooms.repository';
 import { ContentPolicy, JoinPolicy } from '@entities/room.entity';
 import { CreateRoomParams } from '@entities/rooms.repository';
-import { TestRoomsRepository } from '@fixtures/data/test.rooms.repository';
+import { TestRoomsRepository } from '@data/repositories/test/test.rooms.repository';
 import { Test } from '@nestjs/testing';
 import { DataModule } from '../data.module';
 import { MockLoggerModule } from '@fixtures/MockLoggerModule';
