+++ services/api/src/data/repositories/dynamodb/dynamodb.messages.repository.ts
@@ -2,8 +2,8 @@ import { DraftMessage, SentMessage } from '@entities/messages';
 import { MessagesRepository } from '@entities/messages';
 import { Injectable } from '@nestjs/common';
 import { pick } from 'rambda';
-import { DynamoDBAdapter } from '../adapters/dynamodb.adapter';
-import { DbMessage } from '../adapters/schema';
+import { DynamoDBAdapter } from '../../adapters/dynamodb.adapter';
+import { DbMessage } from '../../adapters/schema';
 
 @Injectable()
 export class DynamoDBMessagesRepository extends MessagesRepository {
