+++ services/api/src/data/repositories/dynamodb/dynamodb.users.repository.ts
@@ -6,8 +6,8 @@ import {
 } from '@entities/users';
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { pick } from 'rambda';
-import { DynamoDBAdapter } from '../adapters/dynamodb.adapter';
-import { DbUser } from '../adapters/schema';
+import { DynamoDBAdapter } from '../../adapters/dynamodb.adapter';
+import { DbUser } from '../../adapters/schema';
 
 @Injectable()
 export class DynamoDBUsersRepository extends UsersRepository {
