+++ services/api/src/domain/usecases/memberships/invite.spec.ts
@@ -1,7 +1,7 @@
 import { JoinPolicy } from '@entities/room.entity';
 import { TestAuthService } from '@fixtures/auth/test-auth-service';
 import { TestMembershipsRepository } from '@fixtures/data/test.memberships.repository';
-import { TestRoomsRepository } from '@fixtures/data/test.rooms.repository';
+import { TestRoomsRepository } from '@data/repositories/test/test.rooms.repository';
 import { RoomFactory } from '@fixtures/messages/room.factory';
 import { UserFactory } from '@fixtures/messages/user.factory';
 import { UnauthorizedException } from '@nestjs/common';
@@ -10,7 +10,7 @@ import { Role } from '@usecases/auth.service';
 import mock, { MockProxy } from 'jest-mock-extended/lib/Mock';
 import { Dispatcher } from '@entities/messages/message';
 import { InviteUseCase } from './invite';
-import { TestUsersRepository } from '@fixtures/data/test.users.repository';
+import { TestUsersRepository } from '@data/repositories/test/test.users.repository';
 import { MembershipStatus } from '@entities/membership.entity';
 
 describe('InviteUseCase', () => {
