+++ services/api/src/domain/usecases/messages/send.spec.ts
@@ -1,5 +1,5 @@
 import { UserFactory } from '@fixtures/messages/user.factory';
-import { TestRoomsRepository } from '@fixtures/data/test.rooms.repository';
+import { TestRoomsRepository } from '@data/repositories/test/test.rooms.repository';
 import { RoomFactory } from '@fixtures/messages/room.factory';
 import { TestAuthService } from '@fixtures/auth/test-auth-service';
 import { UnauthorizedException } from '@nestjs/common';
