+++ services/api/src/domain/usecases/rooms/about-room.spec.ts
@@ -1,5 +1,5 @@
 import { TestAuthService } from '@fixtures/auth/test-auth-service';
-import { TestRoomsRepository } from '@fixtures/data/test.rooms.repository';
+import { TestRoomsRepository } from '@data/repositories/test/test.rooms.repository';
 import { RoomFactory } from '@fixtures/messages/room.factory';
 import { UserFactory } from '@fixtures/messages/user.factory';
 import { UnauthorizedException } from '@nestjs/common';
@@ -8,7 +8,7 @@ import { Role } from '@usecases/auth.service';
 import mock, { MockProxy } from 'jest-mock-extended/lib/Mock';
 import { Dispatcher } from '@entities/messages/message';
 import { AboutRoomUseCase } from './about-room';
-import { TestUsersRepository } from '@fixtures/data/test.users.repository';
+import { TestUsersRepository } from '@data/repositories/test/test.users.repository';
 
 describe('AboutRoomUseCase', () => {
   let aboutRoom: AboutRoomUseCase;
