+++ services/api/src/domain/usecases/rooms/create.spec.ts
@@ -1,5 +1,5 @@
 import { TestMembershipsRepository } from '@fixtures/data/test.memberships.repository';
-import { TestRoomsRepository } from '@fixtures/data/test.rooms.repository';
+import { TestRoomsRepository } from '@data/repositories/test/test.rooms.repository';
 import { UserFactory } from '@fixtures/messages/user.factory';
 import { CreateRoomUseCase } from './create';
 import { ContentPolicy, JoinPolicy } from '@entities/room.entity';
