+++ services/api/src/domain/usecases/rooms/get.spec.ts
@@ -1,5 +1,5 @@
 import { TestAuthService } from '@fixtures/auth/test-auth-service';
-import { TestRoomsRepository } from '@fixtures/data/test.rooms.repository';
+import { TestRoomsRepository } from '@data/repositories/test/test.rooms.repository';
 import { RoomFactory } from '@fixtures/messages/room.factory';
 import { UserFactory } from '@fixtures/messages/user.factory';
 import { GetRoomUseCase } from './get';
